+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/boot/BootAccessManagement.java
@@ -22,8 +22,7 @@ import uk.gov.hmcts.reform.roleassignment.config.SecurityConfiguration;
 import uk.gov.hmcts.reform.roleassignment.config.SwaggerConfiguration;
 import uk.gov.hmcts.reform.roleassignment.util.Swagger2SpringBoot;
 import uk.gov.hmcts.rse.ccd.lib.YamlPropertySourceFactory;
-import uk.gov.hmcts.rse.ccd.lib.common.Common;
-import uk.gov.hmcts.rse.ccd.lib.common.DBWaiter;
+import uk.gov.hmcts.rse.ccd.lib.injected.Common;
 
 @SpringBootConfiguration
 @EnableAutoConfiguration
