--- rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/common/Common.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.rse.ccd.lib.common;
-
-public interface Common {
-}
