+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/injected/Common.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.rse.ccd.lib.injected;
+
+public interface Common {
+}
