+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/injected/DBWaiter.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.rse.ccd.lib.common;
+package uk.gov.hmcts.rse.ccd.lib.injected;
 
 import javax.sql.DataSource;
 import lombok.SneakyThrows;
